+++ lib/src/properties/bottom_sheet_props.dart
@@ -14,7 +14,7 @@ class BottomSheetProps {
   final RouteTransitionsBuilder? transitionBuilder;
   final ShapeBorder? shape;
   final bool useRootNavigator;
-  final BoxConstraints constraints;
+  final BoxConstraints? constraints;
   final double? elevation;
   final String? semanticLabel;
   final Color? barrierColor;
@@ -46,7 +46,7 @@ class BottomSheetProps {
     this.actionsOverflowDirection,
     this.clipBehavior = Clip.none,
     this.useRootNavigator = false,
-    this.constraints = const BoxConstraints(maxHeight: 500),
+    this.constraints,
     this.actionsPadding = EdgeInsets.zero,
     this.insetPadding = const EdgeInsets.symmetric(
       horizontal: 40.0,