+++ lib/src/properties/dialog_props.dart
@@ -14,7 +14,6 @@ class DialogProps {
   final RouteTransitionsBuilder? transitionBuilder;
   final ShapeBorder? shape;
   final bool useRootNavigator;
-  final BoxConstraints constraints;
   final double? elevation;
   final String? semanticLabel;
   final Color? barrierColor;
@@ -46,11 +45,6 @@ class DialogProps {
     this.actionsOverflowDirection,
     this.clipBehavior = Clip.none,
     this.useRootNavigator = false,
-    this.constraints = const BoxConstraints(
-      minWidth: 500,
-      maxWidth: 500,
-      maxHeight: 600,
-    ),
     this.actionsPadding = EdgeInsets.zero,
     this.insetPadding = const EdgeInsets.symmetric(
       horizontal: 40.0,