+++ lib/src/properties/menu_props.dart
@@ -4,7 +4,6 @@ import '../../dropdown_search.dart';
 
 class MenuProps {
   final ShapeBorder? shape;
-  final BoxConstraints constraints;
   final double? elevation;
   final Color? barrierColor;
   final Color? backgroundColor;
@@ -31,7 +30,6 @@ class MenuProps {
     this.barrierDismissible = true,
     this.animation,
     this.clipBehavior = Clip.none,
-    this.constraints = const BoxConstraints(maxHeight: 350),
     this.animationDuration = const Duration(milliseconds: 300),
     this.textStyle,
     this.borderOnForeground = false,